% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3.billing.R
\name{billing}
\alias{billing}
\title{Returns the list of all billing groups you have access to}
\usage{
billing(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all billing groups you have access to.
This is an utility method used only create projects with
appropriate billing group. Full access to billing data is not
available via the SBG public API yet.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = billing(token)}
}

