% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.billing.R
\name{billing_groups}
\alias{billing_groups}
\title{[v2] Get a single billing group}
\usage{
billing_groups(auth_token = NULL, id = NULL, breakdown = FALSE, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{id}{billing group id}

\item{breakdown}{This call returns a breakdown of spending
per-project for the billing group specified by specified billing
group id.}
}
\description{
This call list all your billing groups without id provided, and
will retrieves a single billing group, specified by {id}.
}
\details{
Billing periods:The spending will be shown for the current billing
period, unless the billing group specified is a free billing group
(i.e. pilot funds that you are using the CGC with), in which case
all spending on the account will be shown.
}
\examples{
\donttest{
## list all groups, to check id and details
billing_groups(token, base_url = url)
## check specific billing group by provide id
billing_groups(token, base_url = url, id = "2b255c92-2b54-4fc8-something")
## breakdown specific group
billing_groups(token, base_url = url, id = "2b255c92-2b54-4fc8-something", breakdown = TRUE)
}
}

