% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.billing.R
\name{billing_invoices}
\alias{billing_invoices}
\title{List invoices}
\usage{
billing_invoices(auth_token = NULL, id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token.}

\item{id}{invoices id. by default is NULL, will list all invoices
with ID. if id is provided, This call retrieves information about a
selected invoice, including the costs for analysis and storage, and
the invoice period.}
}
\description{
This call returns a list of invoices, with information about each,
including whether or not the invoice is pending and the billing
period it covers. The call returns information about all available
invoices, unless you use the query parameter bg_id to specify the
ID of specific invoices.
}
\examples{
\donttest{
## list all invoices
billing_invoices(token, base_url = url)
## check single invoice
billing_invoices(token, base_url = url, id = "some_id")
}
}

