% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99.easey_api.R
\docType{class}
\name{Metadata-class}
\alias{Metadata}
\alias{Metadata-class}
\title{Metadata class}
\description{
Metadata class
}
\details{
Used for SBG GUI, fixed set of enums and fields are visible for now
}

