% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99.easy_api.R
\docType{class}
\name{FileTypeSingleEnum-class}
\alias{FileTypeSingleEnum-class}
\alias{Metadata}
\alias{Metadata-class}
\alias{file_type}
\alias{paired_end}
\alias{qual_scale}
\alias{seq_tech}
\title{Metadata class}
\description{
Metadata class
}
\details{
This function will help you create a Metadata object, what it does
it to accept a named list or just pass meta key-value pairs as
argument one by one. Then it first matches SBG's build-in meta field
which will be shown in the graphic interface on the platform, then
save extra meta information in extra field, but not visible on the
platform yet, you can view it via the API.

There are four pre-defined fields with pre-defined levels, they are
file_type, qual_scale, seq_tech, and paired_end, those are also
constructor names to construct a single Enum object, it's different
from characters, it has validation against levels, to check their
levels, you can simply create a empty Metadata object and access
the field levels. Please see examples.
}
\examples{
m <- Metadata()
## to check levels
m$file_type
levels(m$file_type)
## to replace a Enum class need to use constructor
m$file_type <- file_type("text")
}
\references{
\url{https://docs.sbgenomics.com/display/sbg/Metadata}
}

