% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/7.misc.R
\name{misc_get_auth_token}
\alias{misc_get_auth_token}
\title{Opens browser to copy the auth token}
\usage{
misc_get_auth_token()
}
\value{
auth token
}
\description{
Click 'Generate Token' button, copy and paste the generated token
string to the R console. The function will return the token string.
}
\examples{
\donttest{misc_get_auth_token()}
}

