% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/7.misc.R
\name{misc_get_uploader}
\alias{misc_get_uploader}
\title{Download SBG uploader and extract to a specified directory}
\usage{
misc_get_uploader(destdir = NULL)
}
\arguments{
\item{destdir}{The directory to extract SBG uploader to.
If not present, it will be created automatically.}
}
\description{
Download SBG uploader and extract to a specified directory.
}
\examples{
\donttest{misc_get_uploader('~/sbg-uploader/')}
}

