% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/7.misc.R
\name{misc_upload_cli}
\alias{misc_upload_cli}
\title{Upload files using SBG uploader}
\usage{
misc_upload_cli(auth_token = NULL, uploader = NULL, file = NULL,
  project_id = NULL, proxy = NULL)
}
\arguments{
\item{auth_token}{auth token}

\item{uploader}{The directory where the SBG uploader is located
(the directory that contains the bin/ directory).}

\item{file}{The location of the file to upload.}

\item{project_id}{The project ID to upload the files to.
If you do not supply this, then the uploader will place the
incoming files in your "My Files" section.}

\item{proxy}{Allows you to specify a proxy server through which
the uploader should connect. About the details the proxy parameter format,
see \url{https://developer.sbgenomics.com/tools/uploader/documentation}.}
}
\description{
Upload files using SBG uploader.
}
\examples{
\donttest{misc_upload_cli(auth_token = 'your token',
                          uploader = '~/sbg-uploader/',
                          file = '~/test.fastq', project_id = '1234')}
}
\references{
\url{https://developer.sbgenomics.com/tools/uploader/documentation}
}

