% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5.pipelines.R
\name{pipeline_list_my}
\alias{pipeline_list_my}
\title{Returns the list of pipelines in user's "My Pipelines" section}
\usage{
pipeline_list_my(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of pipelines in user's "My Pipelines" section.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = pipeline_list_my(token)}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/get/pipeline/my}
}

