% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5.pipelines.R
\name{pipeline_list_pub}
\alias{pipeline_list_pub}
\title{Returns the list of all public pipelines}
\usage{
pipeline_list_pub(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all public pipelines.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = pipeline_list_pub(token)}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/get/pipeline/public}
}

