% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2.projects.R
\name{project_delete}
\alias{project_delete}
\title{Delete a project}
\usage{
project_delete(auth_token = NULL, project_id = NULL)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to delete.}
}
\value{
parsed list of the returned json
}
\description{
Note that this deletes all files, tasks which belong to a project.
}
\examples{
\donttest{token = '410b4672ebfc43bab48dd0d18a32fb6f'
req = project_delete(token, project_id = '3a21ade8-ef3e-41f8-8ac2-1dc3b434ac77')}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/delete/project/\%3Aproject_id}
}

