% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2.projects.R
\name{project_details}
\alias{project_details}
\title{Returns the details of the project with ID specified by the project_id path parameter}
\usage{
project_details(auth_token = NULL, project_id = NULL)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}
}
\value{
parsed list of the returned json
}
\description{
Returns the details of the project with ID specified by the project_id path parameter.
}
\examples{
\donttest{token = '420b4672ebfc43bab48dc0d18a32fb6f'
req = project_details(token, project_id = 'b0b3a611-6bb0-47e5-add7-a83402cf7858')}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/get/project/\%3Aproject_id}
}

