% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2.projects.R
\name{project_list}
\alias{project_list}
\title{Returns the list of all projects you have access to}
\usage{
project_list(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all projects you have access to.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = project_list(token)}
}

