% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.ratelimit.R
\name{rate_limit}
\alias{rate_limit}
\title{Rate Limit (V2)}
\usage{
rate_limit(auth_token = NULL, ...)
}
\arguments{
\item{auth_token}{auth token.}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
Rate Limit
}
\details{
The following API call allows you to get your rate limit. This call
returns information about your current rate limit. This is the
number of API calls you can make in one hour. Note: that making
this API call does not count against your rate limit.
}
\examples{
\donttest{
req <- rate_limit(token)
}
}

