% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0.core.R
\name{sbgapi}
\alias{sbgapi}
\title{wrapper for http logic}
\usage{
sbgapi(auth_token = NULL, version = "1.1", path, method = c("GET", "POST",
  "PUT", "DELETE"), query = NULL, body = NULL)
}
\value{
returned request list of httr
}
\description{
wrapper of http logic for SBG API
}
\keyword{internal}

