% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgr-package.R
\docType{package}
\name{sbgr-package}
\alias{sbgr-package}
\title{R Client for Seven Bridges Genomics API (v1)}
\description{
R Client for Seven Bridges Genomics API (v1)
}
\details{
The vignette can be opened with \code{vignette('sbgr')}.

\tabular{ll}{ Package: \tab sbgr\cr Type: \tab Package\cr
License: \tab Apache License 2.0\cr }
}
\author{
Nan Xiao <\email{nan.xiao@sevenbridges.com}>
        Teng-Fei Yin <\email{tengfei.yin@sevenbridges.com}>
}

