% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6.tasks.R
\name{task_list}
\alias{task_list}
\title{Returns the list of all the tasks for a project}
\usage{
task_list(auth_token = NULL, project_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{project_id}{ID of a project you want to access.}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all the tasks for a project.
}
\details{
This function returns general information and status of a task,
in case you want to get a details, including the inputs, outputs
and parameters set for that task, you will have to use task details
resource referencing the \code{task_id} of a task that you want to
get information about.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = task_list(token,
                '1c1d06d2-5862-48f6-b595-e0099b20937e')}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/get/project/\%3Aproject_id/task}
}

