% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1.upload.R
\name{upload_delete}
\alias{upload_delete}
\title{Aborts the upload}
\usage{
upload_delete(auth_token = NULL, upload_id = NULL, ...)
}
\arguments{
\item{auth_token}{auth token}

\item{upload_id}{ID of the upload}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
All upload records and the file are deleted.
}
\examples{
token = '420b4672ebfc43bab48dc0d18a32fb6f'
\donttest{req = upload_delete(token,
            upload_id = '8D7sQJxQk14ubsEnKaoeQZlRvV6ouQtMzBWaQNJdxPDLypUC3WogwtJdncevHxnT')}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/delete/upload/multipart/\%3Aupload_id}
}

