% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.user.R
\name{user_list}
\alias{user_list}
\title{List user resources (V2)}
\usage{
user_list(auth_token = NULL, username = NULL, ...)
}
\arguments{
\item{auth_token}{auth token.}

\item{username}{NULL or character.}

\item{...}{parameters passed to sbgapi function}
}
\value{
parsed list of the returned json
}
\description{
List user resources
}
\details{
This call returns a list of the resources, such as projects,
billing groups, and organizations, that are accessible to you.

The call will only return a successful response if {username} is
replaced with your own username, unless you are an
administrator. If you are an administrator, you can replace
{username} with the username of any CGC user, to return information
on their resources.
}
\examples{
\donttest{
req <- user_list(token)
req <- user_list(token, username = "test")
}
}

