% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99.easey_api.R
\docType{class}
\name{Auth-class}
\alias{Auth}
\alias{Auth-class}
\title{Class Auth}
\arguments{
\item{auth_token}{[character] your auth token.}

\item{api}{[character %in% 'sbg-us', 'cgc'] which api you are
using, by default it is sbg us platform.}

\item{url}{[chracter] a URL for the API, default is \code{NULL},
will use \code{api} parameter to switch to the right one.}

\item{version}{[character] default: 1.1 version used for api.}
}
\description{
Auth token object
}
\details{
Every object coud be requested from this Auth object and any action
could start from this object using cascading.
}
\section{Fields}{

\describe{
\item{\code{auth_token}}{[character] your auth token.}

\item{\code{url}}{[character] basic url used for API.}
}}
\section{Methods}{

\describe{
\item{\code{project(name = NULL, id = NULL, index = NULL, ignore.case = TRUE,
  exact = TRUE)}}{find project}
}}

