% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99.easy_api.R
\docType{class}
\name{Item-class}
\alias{Item}
\alias{Item-class}
\title{Class Item}
\description{
Class Item
}
\details{
To describe a set of objects, Project, Task, Pipeline, File etc.
}
\section{Fields}{

\describe{
\item{\code{response}}{save the raw response from a request.}

\item{\code{auth_token}}{propagate the auth_token from parent.}
}}

