% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99.easey_api.R
\docType{class}
\name{Metadata-class}
\alias{Metadata}
\alias{Metadata-class}
\title{Metadata class}
\description{
Metadata class
}
\details{
This funcion will help you create a Metadata object, what it does it to accept
a named list or just pass meta as argument one by one. Then it first match SBG's
build-in meta field which will be shown in the graphic interface on the platform,
then save extra meta infomation in extra field, but not visible on the platform yet,
you can view it via the API.

There are four pre-defined field with pre-defined levels, they are file_type,
qual_scale, seq_tech, and paired_end, to check there levels, you can simply
create a empty Metadata object and access the field levels. Please see examples.

Metadata is in progress, we will keep it updated.
}
\examples{
Metadata()
## to check levels
Metadata()$file_type
levels(Metadata()$file_type)
}

