% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2.projects.R
\name{project_list}
\alias{project_list}
\title{Returns the list of all projects you have access to}
\usage{
project_list(auth_token = NULL)
}
\arguments{
\item{auth_token}{auth token}
}
\value{
parsed list of the returned json
}
\description{
Returns the list of all projects you have access to.
}
\examples{
\donttest{token = '410b4672ebfc43bab48dd0d18a32fb6f'
req = project_list(token)}
}
\references{
\url{https://developer.sbgenomics.com/api/1.1/get/project}
}

