% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.core.R
\name{sbgapi}
\alias{sbgapi}
\title{wrapper of http logic for SBG API}
\usage{
sbgapi(auth_token = NULL, version = "1.1", path, method = c("GET", "POST",
  "PUT", "DELETE"), query = NULL, body = list(),
  base_url = paste0("https://api.sbgenomics.com/", version, "/"))
}
\arguments{
\item{auth_token}{authenticate token string.}

\item{version}{API version number, default 1.1.}

\item{path}{path connected with base_url.}

\item{method}{one of 'GET', 'POST', 'PUT', 'Delete'}

\item{query}{Passed to httr package GET/POST call.}

\item{body}{Passed to httr package GET/POST/PUT/DELETE call.}

\item{base_url}{defeault is 'https://api.sbgenomics.com/1.1'}
}
\value{
returned request list of httr
}
\description{
wrapper of http logic for SBG API
}
\details{
Used for advanced users and the core method for higher level API in
this package, please refer to the easy api manual and the two
vignettes pages for more convenient usage.
}
\examples{
token <- "fake_token"
\donttest{
## list projects
sbgapi(auth_token = token, path = 'project', method = "GET")
}
}
\references{
\url{https://docs.sbgenomics.com/display/developerhub/API}
}

