% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sbgr-package.R
\docType{package}
\name{sbgr-package}
\alias{sbgr-package}
\title{R Interface for Seven Bridges Genomics API}
\description{
R Interface for Seven Bridges Genomics API
}
\details{
The vignette can be opened with \code{vignette('sbgr')}.

\tabular{ll}{ Package: \tab sbgr\cr Type: \tab Package\cr
License: \tab MIT\cr }
}
\author{
Nan Xiao <\email{nanx@uchicago.edu}>
        Teng-Fei Yin <\email{tengfei.yin@sbgenomics.com}>
}

