% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sbgr-package.R
\docType{package}
\name{sbgr-package}
\alias{sbgr-package}
\title{R Interface for Seven Bridges Genomics API}
\description{
R Interface for Seven Bridges Genomics API
}
\details{
The comprehensive user guide can be opened with
\code{vignette('sbgr')}.

\tabular{ll}{ Package: \tab sbgr\cr Type: \tab Package\cr
Version: \tab 1.0\cr License: \tab MIT\cr }
}
\author{
Nan Xiao <\email{address@sbgenomics.com}>
        Teng-Fei Yin <\email{tengfei.yin@sbgenomics.com}>
}

