% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99.easey_api.R
\docType{class}
\name{Task-class}
\alias{Task}
\alias{Task-class}
\title{Task class}
\description{
Task class
}
\section{Fields}{

\describe{
\item{\code{id}}{[characterORNULL] The task ID number, used when referring to the
task in other Seven Bridges API calls}

\item{\code{name}}{[characterORNULL] Name of the task you wish to execute. If this
is not specified, the task will be named automatically.}

\item{\code{description}}{[characterORNULL] Description of the task you wish to execute.}

\item{\code{pipeline_id}}{[characterORNULL] ID of the pipeline you wish to execute.}

\item{\code{pipeline_revision}}{[characterORNULL] Revision number of the pipeline
you wish to execute. If this is not specified, the latest pipeline revision
is used.}

\item{\code{start_time}}{[numericORNULL] start time.}

\item{\code{status}}{[characterORNULL] 1) active: task is currently running.
2) completed: task has finished successfully. 3) aborted: task was aborted by
user. 4) failed: task has failed to finish due to either bad inputs and/or
parameters, or because of the internal infrastructure failures.}

\item{\code{message}}{[characterORNULL] task message}

\item{\code{jobs_completed}}{[numericORNULL] completed jobs}

\item{\code{jobs_total}}{[numericORNULL] total jobs.}

\item{\code{inputs}}{[listORNULL] required for task execution. List of key-value
pairs containing mappings of pipeline input node ID to file IDs.
 Note that you must supply an array of file IDs for each input nodes,
 even if the array is empty.}

\item{\code{parameters}}{[listORNULL] required for task execution. List of key-value
pairs containing mappings of node IDs to apps specific parameters. Note that
you must supply some value for parameters, even if this an empty list of
key-value pairs.}

\item{\code{project_id}}{[characterORNULL] required for task execution.
ID of the project you want to execute the task in.}
}}

